Move ChatAddonManager out of trade/
